Split ReportMerger into single Classes
